Refactor Real robot: Extract vesc class from speed  controller
