+++ robot/real/SpeedController.py
@@ -1,18 +1,16 @@
 import time
 
-import pyvesc
-import serial
-
 from robot.Component import Component
 
 
 class SpeedController(Component):
-    def __init__(self, enable=True):
+    def __init__(self, vesc, enable=True):
+        self.vesc = vesc
         self.enabled = enable
         if not enable:
             print("SPEED CONTROLLER NOT ENABLED, TEST MODE")
-        # TODO find real max and min values
-        self.max_speed = 50000  # In duty cycle mode, this could go up to 95000 (95%), or even a little more, close to 100%
+        self.max_speed = 50000  # In duty cycle mode, this could go up to 95000 (95%), or even a little more,
+        # close to 100%
         self.min_speed = 7000
         self.DELAY_BETWEEN_EXECUTIONS = 0.05  # Delay in seconds between two executions of speed update
         self.BRAKE_CURRENT = 15000  # Break current in mA
@@ -25,8 +23,6 @@ class SpeedController(Component):
         self.min_deceleration_step = 0
         self.acceleration_step = self.max_acceleration_step
         self.deceleration_step = self.max_deceleration_step
-        if enable:
-            self.serial = serial.Serial('/dev/ttyACM0', baudrate=115200, timeout=0.05)
         self.last_execution_time = 0.
         self.tachometer = None
         self.rpm = 0
@@ -65,29 +61,8 @@ class SpeedController(Component):
 
         # If it is time to update
         if (time.time() - self.last_execution_time) > self.DELAY_BETWEEN_EXECUTIONS:
-
             # Update speed
             self.update_speed()
-
-            # Request data from VESC
-            self.serial.write(pyvesc.encode_request(pyvesc.GetValues))
-
-            # Check if new data has been received from VESC
-            if self.serial.in_waiting > 61:
-                (response, consumed) = pyvesc.decode(self.serial.read(61))
-                # print ("Consumed", consumed, "Serial response: ", response)
-                if response is not None:
-                    for sensor, value in response.__dict__.items():
-                        if sensor == 'tachometer':
-                            print("tachometer: ", value)
-                            self.tachometer = value
-                        if sensor == 'rpm':
-                            print("rpm: ", value)
-                            self.rpm = value
-                        if sensor == 'temp_mos1':
-                            print("temp mos1: ", value)
-                        if sensor == 'v_in':
-                            print("v_in: ", value)
 
     def update_speed(self):
         # Compute time elapsed since last update
@@ -110,25 +85,9 @@ class SpeedController(Component):
             # If we have already decelerated to min_speed, set speed to zero
             if self.speed < self.min_speed:
                 self.speed = 0
+
         # Send command to VESC
         self.send_speed_command()
 
-    # Send speed command to VESC
     def send_speed_command(self):
-        print("Sending to VESC. Duty Cycle : %s" % self.speed)
-        # noinspection PyArgumentList
-        msg = pyvesc.SetDutyCycle(int(self.speed))
-        packet = pyvesc.encode(msg)
-        if self.enabled:
-            self.serial.write(packet)
-
-    # Send brake commande to VESC
-    def send_brake_command(self):
-        # noinspection PyArgumentList
-        msg = pyvesc.SetCurrentBrake(self.BRAKE_CURRENT)
-        packet = pyvesc.encode(msg)
-        if self.enabled:
-            self.serial.write(packet)
-
-    def get_tacho(self):
-        return self.tachometer
+        self.vesc.send_speed_command(self.speed)
